+++ flutter_inappwebview_ios/ios/Classes/InAppWebView/InAppWebView.swift
@@ -2780,7 +2780,11 @@ public class InAppWebView: WKWebView, UIScrollViewDelegate, WKUIDelegate,
 //    }
     
     public func userContentController(_ userContentController: WKUserContentController, didReceive message: WKScriptMessage) {
-        if message.name.starts(with: "console") {
+        guard let body = message.body as? [String: Any?] else {
+            return
+        }
+        
+        if ["consoleLog", "consoleDebug", "consoleError", "consoleInfo", "consoleWarn"].contains(message.name) {
             var messageLevel = 1
             switch (message.name) {
                 case "consoleLog":
@@ -2804,8 +2808,7 @@ public class InAppWebView: WKWebView, UIScrollViewDelegate, WKUIDelegate,
                     messageLevel = 1
                     break;
             }
-            let body = message.body as! [String: Any?]
-            let consoleMessage = body["message"] as! String
+            let consoleMessage = body["message"] as? String ?? ""
             
             let _windowId = body["_windowId"] as? Int64
             var webView = self
@@ -2813,10 +2816,7 @@ public class InAppWebView: WKWebView, UIScrollViewDelegate, WKUIDelegate,
                 webView = webViewTransport.webView
             }
             webView.channelDelegate?.onConsoleMessage(message: consoleMessage, messageLevel: messageLevel)
-        } else if message.name == "callHandler" {
-            let body = message.body as! [String: Any?]
-            let handlerName = body["handlerName"] as! String
-            
+        } else if message.name == "callHandler", let handlerName = body["handlerName"] as? String {
             if handlerName == "onPrintRequest" {
                 let settings = PrintJobSettings()
                 settings.handledByClient = true
@@ -2839,8 +2839,8 @@ public class InAppWebView: WKWebView, UIScrollViewDelegate, WKUIDelegate,
                 return
             }
             
-            let _callHandlerID = body["_callHandlerID"] as! Int64
-            let args = body["args"] as! String
+            let _callHandlerID = body["_callHandlerID"] as? Int64 ?? 0
+            let args = body["args"] as? String ?? ""
             
             let _windowId = body["_windowId"] as? Int64
             var webView = self
@@ -2877,12 +2877,11 @@ if(window.\(JAVASCRIPT_BRIDGE_NAME)[\(_callHandlerID)] != null) {
             if let channelDelegate = webView.channelDelegate {
                 channelDelegate.onCallJsHandler(handlerName: handlerName, args: args, callback: callback)
             }
-        } else if message.name == "onFindResultReceived" {
-            let body = message.body as! [String: Any?]
-            let findResult = body["findResult"] as! [String: Any]
-            let activeMatchOrdinal = findResult["activeMatchOrdinal"] as! Int
-            let numberOfMatches = findResult["numberOfMatches"] as! Int
-            let isDoneCounting = findResult["isDoneCounting"] as! Bool
+        } else if message.name == "onFindResultReceived",
+                  let findResult = body["findResult"] as? [String: Any],
+                  let activeMatchOrdinal = findResult["activeMatchOrdinal"] as? Int,
+                  let numberOfMatches = findResult["numberOfMatches"] as? Int,
+                  let isDoneCounting = findResult["isDoneCounting"] as? Bool {
             
             let _windowId = body["_windowId"] as? Int64
             var webView = self
@@ -2891,20 +2890,17 @@ if(window.\(JAVASCRIPT_BRIDGE_NAME)[\(_callHandlerID)] != null) {
             }
             webView.findInteractionController?.channelDelegate?.onFindResultReceived(activeMatchOrdinal: activeMatchOrdinal, numberOfMatches: numberOfMatches, isDoneCounting: isDoneCounting)
             webView.channelDelegate?.onFindResultReceived(activeMatchOrdinal: activeMatchOrdinal, numberOfMatches: numberOfMatches, isDoneCounting: isDoneCounting)
-        } else if message.name == "onCallAsyncJavaScriptResultBelowIOS14Received" {
-            let body = message.body as! [String: Any?]
-            let resultUuid = body["resultUuid"] as! String
-            if let result = callAsyncJavaScriptBelowIOS14Results[resultUuid] {
-                result([
-                        "value": body["value"],
-                        "error": body["error"]
-                ])
-                callAsyncJavaScriptBelowIOS14Results.removeValue(forKey: resultUuid)
-            }
-        } else if message.name == "onWebMessagePortMessageReceived" {
-            let body = message.body as! [String: Any?]
-            let webMessageChannelId = body["webMessageChannelId"] as! String
-            let index = body["index"] as! Int64
+        } else if message.name == "onCallAsyncJavaScriptResultBelowIOS14Received",
+                  let resultUuid = body["resultUuid"] as? String,
+                  let result = callAsyncJavaScriptBelowIOS14Results[resultUuid] {
+            result([
+                    "value": body["value"],
+                    "error": body["error"]
+            ])
+            callAsyncJavaScriptBelowIOS14Results.removeValue(forKey: resultUuid)
+        } else if message.name == "onWebMessagePortMessageReceived",
+                  let webMessageChannelId = body["webMessageChannelId"] as? String,
+                  let index = body["index"] as? Int64 {
             var webMessage: WebMessage? = nil
             if let webMessageMap = body["message"] as? [String : Any?] {
                 webMessage = WebMessage.fromMap(map: webMessageMap)
@@ -2913,9 +2909,7 @@ if(window.\(JAVASCRIPT_BRIDGE_NAME)[\(_callHandlerID)] != null) {
             if let webMessageChannel = webMessageChannels[webMessageChannelId] {
                 webMessageChannel.channelDelegate?.onMessage(index: index, message: webMessage)
             }
-        } else if message.name == "onWebMessageListenerPostMessageReceived" {
-            let body = message.body as! [String: Any?]
-            let jsObjectName = body["jsObjectName"] as! String
+        } else if message.name == "onWebMessageListenerPostMessageReceived", let jsObjectName = body["jsObjectName"] as? String {
             var webMessage: WebMessage? = nil
             if let webMessageMap = body["message"] as? [String : Any?] {
                 webMessage = WebMessage.fromMap(map: webMessageMap)