+++ flutter_inappwebview/example/integration_test/in_app_webview/on_received_icon.dart
@@ -26,7 +26,9 @@ void onReceivedIcon() {
             pageLoaded.complete();
           },
           onReceivedIcon: (controller, icon) {
-            onReceivedIconCompleter.complete(icon);
+            if (!onReceivedIconCompleter.isCompleted) {
+              onReceivedIconCompleter.complete(icon);
+            }
           },
         ),
       ),