+++ flutter_inappwebview/example/lib/in_app_browser_example.screen.dart
@@ -61,6 +61,10 @@ class MyInAppBrowser extends InAppBrowser {
     print("\n\nOverride ${navigationAction.request.url}\n\n");
     return NavigationActionPolicy.ALLOW;
   }
+
+  void onMainWindowWillClose() {
+    close();
+  }
 }
 
 class InAppBrowserExampleScreen extends StatefulWidget {