+++ flutter_inappwebview/lib/src/in_app_browser/in_app_browser.dart
@@ -538,4 +538,7 @@ class InAppBrowser implements PlatformInAppBrowserEvents {
       NavigationAction navigationAction) {
     return null;
   }
+
+  @override
+  void onMainWindowWillClose() {}
 }